refactor(dashboard): extract all tab content into DashboardTabContent
Move the entire AnimatePresence tab region (overview/picks/milestones/
genres — four sequential, contiguous `activeTab === "…"` blocks)
verbatim into _components/dashboard-tab-content.tsx. Component owns its
own Dashboard/Dashboard.body i18n + router (same pattern as the other
extractions); everything else threaded as props (types derived from the
hooks' ReturnTypes). The ~1,167-line JSX is byte-identical (sed-copied)
so render output is identical by construction. Also removed the now-dead
in-page chart dynamic() defs + unused Image/dynamic imports.

Done in one big commit (speed-prioritized per request; the clean
sequential-block shape made it a safe single verbatim move). tsc + build
+ 93 tests green. dashboard/page.tsx 1558->401 LOC (2439->401, -84%
across the sweep).
